refactor(CardItem): move CardItem into ui.component
