+++ app/src/main/java/tcg/pocket/dex/allcards/AllCardScreen.kt
@@ -8,6 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import tcg.pocket.dex.component.CardItem
 import tcg.pocket.dex.tierdecks.fakeCardsData
 import tcg.pocket.dex.ui.theme.TcgPocketDexTheme
