+++ app/src/main/java/tcg/pocket/dex/allcards/RelatedCardsSection.kt
@@ -10,6 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import tcg.pocket.dex.component.CardItem
 import tcg.pocket.dex.component.PocketDexSectionHeader
 import tcg.pocket.dex.tierdecks.fakeCardsData
 import tcg.pocket.dex.ui.theme.TcgPocketDexTheme
